+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 * :issue:`30`: Add :meth:`pymaf.MafFrame.plot_rainfall` method.
 * :issue:`30`: Add :meth:`pyvcf.VcfFrame.plot_rainfall` method.
 * Update :meth:`pymaf.MafFrame.to_vcf` method to output sorted VCF.
+* Add :meth:`pymaf.MafFrame.matrix_prevalence` method.
 
 0.16.0 (2021-07-02)
 -------------------